Refine UI text and refactor loaders
- Localize hardcoded "lines" text on deck check with translation keys
- Clarify search description to mention names, types, and rules text instead of just card names
- Refactor routes into directory structure with separate loader files
- Move route components to route.tsx files in respective directories
- Extract loader functions to dedicated loader.ts files
- Maintain all existing URLs and functionality
